requirements.txt
aiohttp
attrs>=19.3.0
python-box>=3.4.5
requests>=2.22.0
tqdm>=4.38.0
xmltodict>=0.12.0